+++ src/lib/util/chain-util.js
@@ -1,0 +1,9 @@
+function getNodeTriggeringJob(context, nodeChain) {
+  return nodeChain.find(
+    e =>
+      e.project === context.config.github.repository ||
+      e.project === context.config.github.project
+  );
+}
+
+module.exports = { getNodeTriggeringJob };