+++ test/lib/flows/pull-request-flow.test.js
@@ -85,8 +85,8 @@ test("start no parent dependencies. project triggering the job", async () => {
   );
   expect(parentChainFromNode).toHaveBeenCalledWith(definitionTree);
   expect(checkoutDefinitionTree).toHaveBeenCalledWith(context, [
-    { project: "project2" },
-    { project }
+    { project },
+    { project: "project2" }
   ]);
 
   expect(printCheckoutInformation).toHaveBeenCalledTimes(1);
@@ -161,8 +161,8 @@ test("start no parent dependencies. project triggering the job. isArchiveArtifac
   );
   expect(parentChainFromNode).toHaveBeenCalledWith(definitionTree);
   expect(checkoutDefinitionTree).toHaveBeenCalledWith(context, [
-    { project: "project2" },
-    { project }
+    { project },
+    { project: "project2" }
   ]);
 
   expect(printCheckoutInformation).toHaveBeenCalledTimes(1);
@@ -239,8 +239,8 @@ test("start no parent dependencies. project triggering the job. Execute Exceptio
   );
   expect(parentChainFromNode).toHaveBeenCalledWith(definitionTree);
   expect(checkoutDefinitionTree).toHaveBeenCalledWith(context, [
-    { project: "project2" },
-    { project }
+    { project },
+    { project: "project2" }
   ]);
 
   expect(printCheckoutInformation).toHaveBeenCalledTimes(1);